optimize VarDeclaration instance size
VarDeclaration size goes from 305 to 297 byte, saving the allocation of an extra 16 byte block.

+++ src/dmd/declaration.d
@@ -1104,7 +1104,7 @@ extern (C++) class VarDeclaration : Declaration
     bool onstack;                   // it is a class that was allocated on the stack
     bool mynew;                     // it is a class new'd with custom operator new
 
-    int canassign;                  // it can be assigned to
+    byte canassign;                  // it can be assigned to
     bool overlapped;                // if it is a field and has overlapping
     bool overlapUnsafe;             // if it is an overlapping field and the overlaps are unsafe
     bool doNotInferScope;           // do not infer 'scope' for this variable
